Don't overwrite hooks without --force
Currently, `bark install` silently overwrites any
`reference-transaction` hook that may be present. This is a
potentially very destructive action, so we should take proper care
before doing so.

I don't think there's a good way to automatically resolve the conflict
in case the file already exists, so this implementation simply lists
the conflicting files and leaves it to the user to resolve the
conflict manually. (For example, we could move both hook files into
`hooks/reference-transaction.d/` and make
`hooks/reference-transaction` a script that calls all scripts in
`reference-transaction.d/`, but it's not obvious that things like
order of execution wouldn't matter.)